Move process details to separate component
